requirements.txt
amqp==5.0.5
appdirs==1.4.3
asgiref==3.2.5
billiard==3.6.3.0
Brotli==1.0.7
celery==5.0.5
certifi==2019.11.28
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Deprecated==1.2.11
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-kanban-budget==0.1.4
django-uuslug==1.2.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.5.0
kombu==5.0.2
Markdown==3.3.3
Pillow==8.1.0
prompt-toolkit==3.0.16
psycopg2==2.8.4
PyGithub==1.54.1
PyJWT==1.7.1
python-slugify==4.0.1
pytz==2019.3
requests==2.25.1
six==1.14.0
slacker==0.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.26.3
vine==5.0.0
virtualenv==20.0.12
virtualenv-clone==0.5.3
wcwidth==0.2.5
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.1.0
